requirements-base.txt
aiofiles
sqlalchemy-filters
SQLAlchemy-Searchable
alembic
arrow
cachetools==3.1.1  # NOTE pinning for google-auth
click
email-validator
emails
fastapi
google-api-python-client
google-auth-oauthlib
jinja2
jira
joblib
oauth2client
psycopg2-binary
pyparsing
pypd  # pagerduty plugin
python-jose
python-multipart
pytz
requests
schedule
sentry-asgi
sentry_sdk
slackclient
spacy
sqlalchemy
tabulate
tenacity
uvicorn
